match more pika functions

+++ src/melee/ft/chara/ftpikachu.h
@@ -6,11 +6,11 @@
 
 
 typedef struct _ftPikachuAttributes {
-    u32 x0;
-    u32 x4;
-    u32 x8;
-    u32 xC;
-    u32 x10;
+    f32 x0;
+    f32 x4;
+    f32 x8;
+    f32 xC;
+    f32 x10;
     u32 x14;
     u32 x18;
     u32 x1C_padding[48];
